Add script processor
The script processor executes Javascript code to process an event. The processor uses a pure
Go implementation of ECMAScript 5.1. This can be useful in situations where one of the other
processors doesn’t provide the functionality you need to filter events.

The processor can be configured by embedding Javascript in your configuration file or by pointing
the processor at external file(s). See the included documentation for the full details.

    processors:
    - script:
        type: javascript
        code: >
          function process(event) {
              event.Tag("js");
          }

For observability it can add metrics with a histogram of the execution time and a counter
for the number of exceptions that occur.

Benchmark Results

Here's a benchmark to get a rough idea of how long it takes to process an event in the processor's runtime

    goos: darwin
    goarch: amd64
    pkg: github.com/elastic/beats/libbeat/processors/script/javascript
    BenchmarkBeatEventV0/Put-12                         2000000           712 ns/op
    BenchmarkBeatEventV0/Object_Put_Key-12              2000000           615 ns/op
    BenchmarkBeatEventV0/Get-12                         2000000           769 ns/op
    BenchmarkBeatEventV0/Get_Undefined_Key-12           1000000          1010 ns/op
    BenchmarkBeatEventV0/fields_get_key-12              2000000           836 ns/op
    BenchmarkBeatEventV0/Get_@metadata-12               2000000           676 ns/op
    BenchmarkBeatEventV0/Put_@metadata-12               2000000           736 ns/op
    BenchmarkBeatEventV0/Delete_@metadata-12            2000000           651 ns/op
    BenchmarkBeatEventV0/Cancel-12                      3000000           565 ns/op
    BenchmarkBeatEventV0/Tag-12                         2000000           865 ns/op
    BenchmarkBeatEventV0/AppendTo-12                    2000000           630 ns/op