Convert CEF extension value data types
Previously all of cef.extensions.* were string values. Now it uses the data types that are defined for each CEF field to do the translation within the processor. If the field is not the correct data type then the value is dropped from the event in order to avoid mapping exceptions at index time.